requirements.txt
transformers
flask
pytube
pytorch
SpeechRecognition
pydub